hkdf: move custom HKDF-Expand-Label implementation to a separate package
