+++ internal/crypto/hkdf/hkdf_expand_label.go
@@ -1,4 +1,4 @@
-package handshake
+package hkdf
 
 import (
 	"crypto"
@@ -7,10 +7,10 @@ import (
 	"golang.org/x/crypto/hkdf"
 )
 
-// hkdfExpandLabel HKDF expands a label as defined in RFC 8446, section 7.1.
+// ExpandLabel HKDF expands a label as defined in RFC 8446, section 7.1.
 // Since this implementation avoids using a cryptobyte.Builder, it is about 15% faster than the
-// hkdfExpandLabel in the standard library.
-func hkdfExpandLabel(hash crypto.Hash, secret, context []byte, label string, length int) []byte {
+// ExpandLabel in the standard library.
+func ExpandLabel(hash crypto.Hash, secret, context []byte, label string, length int) []byte {
 	b := make([]byte, 3, 3+6+len(label)+1+len(context))
 	binary.BigEndian.PutUint16(b, uint16(length))
 	b[2] = uint8(6 + len(label))
