+++ internal/crypto/hkdf/hkdf_suite_test.go
@@ -1,0 +1,13 @@
+package hkdf_test
+
+import (
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+func TestHkdf(t *testing.T) {
+	RegisterFailHandler(Fail)
+	RunSpecs(t, "HKDF Suite")
+}
