+++ internal/handshake/aead.go
@@ -3,6 +3,7 @@ package handshake
 import (
 	"encoding/binary"
 
+	"github.com/quic-go/quic-go/internal/crypto/hkdf"
 	"github.com/quic-go/quic-go/internal/protocol"
 )
 
@@ -13,8 +14,8 @@ func createAEAD(suite *cipherSuite, trafficSecret []byte, v protocol.Version) *x
 		keyLabel = hkdfLabelKeyV2
 		ivLabel = hkdfLabelIVV2
 	}
-	key := hkdfExpandLabel(suite.Hash, trafficSecret, []byte{}, keyLabel, suite.KeyLen)
-	iv := hkdfExpandLabel(suite.Hash, trafficSecret, []byte{}, ivLabel, suite.IVLen())
+	key := hkdf.ExpandLabel(suite.Hash, trafficSecret, []byte{}, keyLabel, suite.KeyLen)
+	iv := hkdf.ExpandLabel(suite.Hash, trafficSecret, []byte{}, ivLabel, suite.IVLen())
 	return suite.AEAD(key, iv)
 }
