+++ internal/handshake/initial_aead.go
@@ -4,8 +4,9 @@ import (
 	"crypto"
 	"crypto/tls"
 
-	"golang.org/x/crypto/hkdf"
+	stdhkdf "golang.org/x/crypto/hkdf"
 
+	"github.com/quic-go/quic-go/internal/crypto/hkdf"
 	"github.com/quic-go/quic-go/internal/protocol"
 )
 
@@ -52,9 +53,9 @@ func NewInitialAEAD(connID protocol.ConnectionID, pers protocol.Perspective, v p
 }
 
 func computeSecrets(connID protocol.ConnectionID, v protocol.Version) (clientSecret, serverSecret []byte) {
-	initialSecret := hkdf.Extract(crypto.SHA256.New, connID.Bytes(), getSalt(v))
-	clientSecret = hkdfExpandLabel(crypto.SHA256, initialSecret, []byte{}, "client in", crypto.SHA256.Size())
-	serverSecret = hkdfExpandLabel(crypto.SHA256, initialSecret, []byte{}, "server in", crypto.SHA256.Size())
+	initialSecret := stdhkdf.Extract(crypto.SHA256.New, connID.Bytes(), getSalt(v))
+	clientSecret = hkdf.ExpandLabel(crypto.SHA256, initialSecret, []byte{}, "client in", crypto.SHA256.Size())
+	serverSecret = hkdf.ExpandLabel(crypto.SHA256, initialSecret, []byte{}, "server in", crypto.SHA256.Size())
 	return
 }
 
@@ -65,7 +66,7 @@ func computeInitialKeyAndIV(secret []byte, v protocol.Version) (key, iv []byte) 
 		keyLabel = hkdfLabelKeyV2
 		ivLabel = hkdfLabelIVV2
 	}
-	key = hkdfExpandLabel(crypto.SHA256, secret, []byte{}, keyLabel, 16)
-	iv = hkdfExpandLabel(crypto.SHA256, secret, []byte{}, ivLabel, 12)
+	key = hkdf.ExpandLabel(crypto.SHA256, secret, []byte{}, keyLabel, 16)
+	iv = hkdf.ExpandLabel(crypto.SHA256, secret, []byte{}, ivLabel, 12)
 	return
 }
